refactor: use network subcommand for launching local network
