+++ crates/pop-cli/src/commands/up/network.rs
